Speed.

+++ capivara/alphabeta.go
@@ -11,7 +11,7 @@ const (
 )
 
 type alphaBetaState struct {
-	nodes          int
+	nodes          int64
 	showSearch     bool
 	deadline       time.Time
 	cancelled      bool
@@ -50,7 +50,7 @@ func rootAlphaBeta(ab *alphaBetaState, b board, depth int, path []string, addChi
 		child := children[0]
 		score, childPath := alphaBeta(ab, child, -beta, -alpha, depth-1, append(path, child.lastMove), addChildren)
 		score = -score
-		ab.nodes += len(children)
+		ab.nodes += int64(len(children))
 		if ab.showSearch {
 			fmt.Printf("rootAlphaBeta: depth=%d nodes=%d score=%v move: %s path: %s\n", depth, ab.nodes, score, child.lastMove, childPath)
 		}
@@ -76,7 +76,7 @@ func rootAlphaBeta(ab *alphaBetaState, b board, depth int, path []string, addChi
 		}
 		score, childPath := alphaBeta(ab, child, -beta, -alpha, depth-1, append(path, child.lastMove), addChildren)
 		score = -score
-		ab.nodes += len(children)
+		ab.nodes += int64(len(children))
 		if ab.showSearch {
 			fmt.Printf("rootAlphaBeta: depth=%d nodes=%d score=%v move: %s path: %s\n", depth, ab.nodes, score, child.lastMove, childPath)
 		}
@@ -119,7 +119,7 @@ func alphaBeta(ab *alphaBetaState, b board, alpha, beta float32, depth int, path
 		}
 		score, childPath := alphaBeta(ab, child, -beta, -alpha, depth-1, append(path, child.lastMove), addChildren)
 		score = -score
-		ab.nodes += len(children)
+		ab.nodes += int64(len(children))
 		if score >= beta {
 			return beta, childPath
 		}
